+++ CHANGELOG.md
@@ -7,11 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.3.4] - 2022-12-03
 
 ### Changed
+[#31](https://github.com/ssenart/home-assistant-gazpar/issues/31): [Feature] For Weekly readings, provide data on the last 10 weeks VS the same weeks one year before.
+
 [#30](https://github.com/ssenart/home-assistant-gazpar/issues/30): Upgrade PyGazpar to version 1.2.0.
 
-[#27](https://github.com/ssenart/home-assistant-gazpar/issues/27): Energy Dashboard - Unit error - Regression in HA 2022.11.
+[#27](https://github.com/ssenart/home-assistant-gazpar/issues/27): [Issue] Energy Dashboard - Unit error - Regression in HA 2022.11.
 
 ### Fixed
+[#24](https://github.com/ssenart/home-assistant-gazpar/issues/24): [Bug] gas_energy not showing in the energy dashboard
+
 [#28](https://github.com/ssenart/home-assistant-gazpar/issues/28): [Issue] No data update - GrDF web site is half broken - Download button does not work anymore.
 
 ## [1.3.3] - 2022-11-26